PROXY-11: rename module
